Rename to `listAllExams.js`
